Added add and delete friend in user route and fixed cannot send header err

+++ controllers/userController.js
@@ -58,19 +58,59 @@ module.exports = {
 
   async deleteUser(req, res) {
     try {
-      const user = await User.findOneAndDelete(
-        { _id: req.params.userId },
-      );
+      const user = await User.findOneAndDelete({ _id: req.params.userId });
+
       if (!user) {
-        res.status(404).json({ message: "No user found with that id.!" });
+        return res.status(404).json({ message: "No user with that ID" });
       }
 
-      //NOT WORKING FOR SOME REASON>>>>>>>>>>>>>>>>>>>>
-
-      res.status(200).json(user);
+      await Thought.deleteMany({ _id: { $in: user.thoughts } });
+      return res.status(200).json({
+        message: "User and associated thoughts and reactions deleted!",
+      });
     } catch (err) {
+      console.log(err);
+      return res.status(500).json(err);
+    }
+  },
+
+  //Add friend
+  async addFriend(req, res) {
+    try {
+      const friend = await User.findOneAndUpdate(
+        { _id: req.params.userId },
+        { $addToSet: { friends: req.params.friendId } },
+        { runValidators: true, new: true }
+      );
+
+      if (!friend) {
+        res.status(404).json({ message: "No user with that ID" });
+      }
+
+      res.status(200).json(friend);
+    } catch (err) {
+      console.log(err);
       res.status(500).json(err);
     }
-}
+  },
 
-}
+  //Delete friend
+  async deleteFriend(req, res) {
+    try {
+      const friend = await User.findOneAndUpdate(
+        { _id: req.params.userId },
+        { $pull: { friends: req.params.friendId } },
+        { runValidators: true, new: true }
+      );
+
+      if (!friend) {
+        return res.status(404).json({ message: "Check user and friend ID" });
+      }
+
+      return res.status(200).json(friend);
+    } catch (err) {
+      console.log(err);
+      return res.status(500).json(err);
+    }
+  },
+};
